Move health related classes to own dir
