+++ game/src/ecs/components/ai/transition/SelfDefendTransition.java
@@ -1,7 +1,7 @@
 package ecs.components.ai.transition;
 
-import ecs.components.HealthComponent;
 import ecs.components.MissingComponentException;
+import ecs.components.health.HealthComponent;
 import ecs.entities.Entity;
 
 public class SelfDefendTransition implements ITransition {
