+++ game/src/ecs/components/health/death/RadiusDamage.java
@@ -1,0 +1,9 @@
+package ecs.components.health.death;
+
+import ecs.entities.Entity;
+
+public class RadiusDamage implements IOnDeathFunction {
+
+    @Override
+    public void onDeath(Entity entity) {}
+}
